修改了注释错误，修正了删除文件夹所有文件方法中，会把指定的源文件夹都删除的 BUG

+++ generator/helper.js
@@ -142,8 +142,8 @@ function sortArticleByTime(articleList) {
 }
 
 /**
- * 清晰文章元数据
- * 当前清晰功能
+ * 清洗文章元数据
+ * 当前清洗功能
  * 1. 按照时间对文章列表排序，最新的文章排最前
  * @param articleMeta 文章元数据
  */
@@ -169,20 +169,23 @@ function forInArticleList(articleMeta, func) {
 /**
  * 删除指定文件夹下所有文件
  * @param path 要删除文件夹的路径
+ * @param countLevel 文件夹层级
  */
-function removeDirectoryFile(path) {
+function removeDirectoryFile(path, countLevel=1) {
   if (fs.existsSync(path)) {
     const files = fs.readdirSync(path);
     files.forEach(fileName => {
       const currFilePath = `${path}/${fileName}`;
       if (fs.statSync(currFilePath).isDirectory()) {
-        removeDirectoryFile(currFilePath);
+        removeDirectoryFile(currFilePath, countLevel + 1);
       } else {
         fs.unlinkSync(currFilePath);
       }
     });
 
-    fs.rmdirSync(path);
+    if (countLevel !== 1) {
+      fs.rmdirSync(path);
+    }
   }
 }
